requirements.txt
Django==3.2.10
djangorestframework==3.12.4
plotly==5.3.1
mysql-connector-python==8.0.28
django-import-export==2.7.0
pandas
